hadoop_stack/recommender_api/pyproject.toml
[tool.poetry]
name = "recommender_api"
version = "0.1.0"
description = ""
authors = ["Paweł Pankiewicz <239565@edu.p.lodz.pl>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.78.0"
pyspark = "^3.2.1"
uvicorn = "^0.17.6"
python-decouple = "^3.6"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
mypy = "^0.960"

[tool.poetry.scripts]
start_api_server="recommender_api.api:run"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
